library/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
alembic>=1.7.0
psycopg2-binary>=2.9.1
pytest>=6.2.5
httpx>=0.18.2
python-dotenv>=0.19.0 